requirements-dev.txt
-r requirements.txt
pytest>=8.0
pytest-django>=4.8
ruff>=0.5
mypy>=1.10
django-stubs>=5.0
